side list

+++ studybud/base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render,redirect
 from django.contrib import messages
+from django.http import HttpResponse
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate,login,logout
 from django.db.models import Q
 from .models import Room,Topic
@@ -26,6 +28,8 @@ def room(request,pk):
     return render(request,'base/room.html',context)
 
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -36,25 +40,42 @@ def createRoom(request):
     context = {'form': form}
     return render(request,'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
-    if request.method == 'POST':
-        form = RoomForm(request.POST , instance = room)
-        if form.is_valid():
-            form.save()
-            return redirect('home')
-    context={'form': form}
-    return render(request,'base/room_form.html', context)
+    form.host = request.host
+    if request.user == room.host:
+        if request.method == 'POST':
+            form = RoomForm(request.POST , instance = room )
+            if form.is_valid():
+                form.save()
+                return redirect('home')
+        context={'form': form}
+        return render(request,'base/room_form.html', context)
+    else:
+        return HttpResponse('You are not allowed')
+
+
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
-    room = Room.objects.get(id=pk)
-    if request.method == 'POST':
-        room.delete()
-        return redirect('home')
+    room = Room.objects.get(id=pk) 
+    if request.user == room.host:
+        
+        if request.method == 'POST':
+            room.delete()
+            return redirect('home')
+        return render(request,'base/delete.html',{'obj':room})
+    else:
+        return HttpResponse('You are not allowed')
 
-    return render(request,'base/delete.html',{'obj':room})
+
+
 def loginPage(request):
-
+    if request.user.is_authenticated():
+        return render('home')
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -73,4 +94,12 @@ def loginPage(request):
             messages.error(request,'wrong Password Or username')
     context={}        
     return render(request,'base/login_register.html',context)
+
+
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
